requirements.txt
python-telegram-bot
moviepy